# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PLN 5S API",
        "contact": {
            "name": "PLN 5S",
            "email": "rafiadji13@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/area/getjenisarea": {
            "get": {
                "tags": [
                    "Area"
                ],
                "summary": "Get Jenis Area",
                "description": "Untuk mendapatkan semua data Jenis Area",
                "operationId": "getJenisArea",
                "responses": {
                    "200": {
                        "description": "Return Jenis Area",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 0,
                                        "jenis": ""
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/area/getareabyjenis/{id}": {
            "get": {
                "tags": [
                    "Area"
                ],
                "summary": "Get Area By Jenis",
                "description": "Untuk mendapatkan semua data Area by jenis",
                "operationId": "getAreaByJenis",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Jenis",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Area List",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 0,
                                        "namaArea": "",
                                        "jenisId": 0,
                                        "jenis": ""
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/nilai/getallpenilaian": {
            "get": {
                "tags": [
                    "Penilaian"
                ],
                "summary": "Get All Penilaian",
                "description": "Untuk mendapatkan semua data Penilaian",
                "operationId": "getAllPenilaian",
                "responses": {
                    "200": {
                        "description": "Return All Data Penilaian",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "penilaian": "Seiri"
                                    },
                                    {
                                        "id": 2,
                                        "penilaian": "Seiton"
                                    },
                                    {
                                        "id": 3,
                                        "penilaian": "Seiso"
                                    },
                                    {
                                        "id": 4,
                                        "penilaian": "Seiketsu"
                                    },
                                    {
                                        "id": 5,
                                        "penilaian": "Shitsuke"
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/tempat/getjenistempat": {
            "get": {
                "tags": [
                    "Tempat"
                ],
                "summary": "Get Jenis Tempat",
                "description": "Untuk mendapatkan semua data Jenis Tempat",
                "operationId": "getJenisTempat",
                "responses": {
                    "200": {
                        "description": "Return Jenis Tempat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 0,
                                        "jenis": ""
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/tempat/gettempatbyjenis/{id}": {
            "get": {
                "tags": [
                    "Tempat"
                ],
                "summary": "Get Tempat By Jenis",
                "description": "Untuk mendapatkan semua data Tempat by jenis",
                "operationId": "getTempatByJenis",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Jenis",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Tempat List",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 0,
                                        "namaTempat": "",
                                        "jenisId": 0,
                                        "jenis": ""
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/tempat/getalltempat": {
            "get": {
                "tags": [
                    "Tempat"
                ],
                "summary": "Get All Tempat",
                "description": "Untuk mendapatkan semua data Tempat",
                "operationId": "getAllTempat",
                "responses": {
                    "200": {
                        "description": "Return All Data Tempat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 0,
                                        "namaTempat": "",
                                        "jenisId": 0,
                                        "jenis": ""
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/user/getalluser": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get All User",
                "description": "Untuk mendapatkan semua data user",
                "operationId": "getAllUser",
                "responses": {
                    "200": {
                        "description": "Return All Data User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 0,
                                        "username": "",
                                        "password": "",
                                        "nama": "",
                                        "role": 0
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/user/getuserbyid/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User By ID",
                "description": "Untuk mendapatkan data user berdasarkan id",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Data User by Id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 0,
                                    "username": "",
                                    "password": "",
                                    "nama": "",
                                    "role": 0
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/getlogin": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "description": "Untuk Login User",
                "operationId": "getLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Masukkan Username",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Masukkan Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return Data User yang login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "berhasil login",
                                    "data": {
                                        "id": 0,
                                        "username": "",
                                        "nama": "",
                                        "role": 0
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register",
                "description": "Untuk Registrasi user",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Masukkan Username",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Masukkan Password",
                                        "type": "string"
                                    },
                                    "nama": {
                                        "description": "Masukkan Nama",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return json info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil disimpan"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}